Fixing Amazon check
Apparently if you make less than $4300 (really?) a month you can only do this search every 60 seconds. Normally that’s not a problem. When we get popular’d by Wynonna Earp (or Tracy’s awesome posts) it causes some ugly errors to show.

In order to check this, a transient is set for 59 seconds every time this is called. Which is so stupid. If it’s less than 59 seconds, then the ad defaults to the fallback.

This. Is. Ridonkulious.