Year Range
Since many shows get a final season announced, it will save us work to be able to pick a +1 year for finale year.

Also changing the start year to match what the first year a show might be from based on history. We have the define FIRST_LWTV_YEAR and should use it.