Related Shows Patch
From @keesiemeijer

Since plugin version 2.6.0 it's allowed to query for post IDs instead of 
post objects in the shortcode (and widget). This speeds up the related 
posts query considerably.

See the ID query documentation.

This PR adds the fields="ids" attribute to the shortcode and updates the 
alter_results() method to both work with post objects and ids.

Fixes #145