Of the Day Hashtag Improvements
In order to properly tag shows like Charmed (which have "Charmed" and 
"Charmed (2018)") and shows that have anime AND drama (and thus get 
names like "Foo (Drama)"), the code for hashtagging of-the-days will now 
OMIT text between parenthesis. This should not have an impact on any 
normal shows, and will keep the hashtag reasonable.

#Charmed