The Tambour Takedown: Checking Queer IRL
This is a little more complicated. We don't want to award points if a 
show has primarily cast a cis/het actor as a queer role. To solve this, 
we grab the actor listed as PRIMARY (i.e. the one listed first). If THEY 
are QIRL, the show gets the points.