Fixing Select 2 Integration
Using Select2 to manage certain taxonomoies and prevent typos is 
awesome. SAVING them is not. It's a mess becuase CMB2 saves them to a 
postmeta, not the taxonomy, which results in us needing to save twice 
and sometimes you get weird blank data and it deletes the postmeta. To 
fix it, the following happened:

* update Select2 library to 4.0.7
* fork CMB2's implimentation and remove from composer
* rewrite the integration to properly saves
* add in proper checks for defaults
* remove unused code
* update cache code