Char years (#180)
Adding support for years for char activity

Support for listing characters by year they were on a specific show. You can thank the complexity to our favorite character, Sara Lance. The years are added per-show, and are repeatable so you have to add each one. This can be really annoying, but will cover gaps in appearances.