Fixing On-Air logic flaw
If a show is pre-set to be off-air then it may show as 'on air' on the 
shows listing when it's not. 

To fix, I've added in a new meta value for 'On Air' that gets updated 
when the show score is updated (which happens on any un-cached pageload 
or admin-visit). If that's been set, then the facet checker uses that.

Also the calucator now updates the facets, which will pick up all 
changes to the scores etc.

Snuck in a failsafe for canceled shows and whats-on-next.